Bug fix, enemy lasers go to the bottom of the screen
Set the canvasWidth and canvasHeight options when creating an enemy, as the lasers stopped travelling to 640 pixels - the default value for canvasHeight when it's not set. So in the Enemy.js file we can remove the hack that ensured the laser would hit the player. Also add 'use strict' to the getRandomInt function for consistency