Project Fix-Game-Bugs-And-Hacks
Previously called Project Make-Game-Easier-And-More-Fun has been renamed for what it really is. Now in phase 2, we address the issue of being able to shoot by frames  as it was always a hack to get the game out. Not really sustainable as framerates can differ from machine to machine. For example, the bug fix in commit 3c99a4acc6978ccbe7f3bba7a6d3650b34bb351f really changed the playability of the game. So we go for a more standardised measurement to allow shots to be fired - time. Every sprite now has a set time which they have to wait before they can shoot again.