Allow user to try again after defeat
Add a reset function for the player and game. Now a 'Try Again?' options comes up so the user doesn't have the refresh the browser to continue.  Do some other fixes such as initialise the enemyOptions variable only if the game is neither paused nor over. Also, convert all font to pt instead of px and fix san-serif to sans-serif fallback in text fonts